requirements.txt
openai-whisper
sounddevice
soundfile
numpy
PySide6
pyannote.audio
python-dotenv
openai
anthropic
pytest
